Rework Apple Position Generation
- Previously: Head picks temporary pos, ask each body part if it's okay,
if it is Head sets the new apple pos else it restarts the loop. This
work but could take several frames to find a valid apple position.
- Now: Head saves all the pos occupied by body parts, then picks a
random unoccupied one. I think this works (not 100% sure, it's 1am) but
I'm not a fan of saving the body positions in both memory and a bunch of
entities, but well the previous code wasn't clean either.